% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{Beta}
\alias{Beta}
\title{Beta}
\usage{
Beta(mat.z, mat.y.ex, n.eq, p, est.mode, iter)
}
\arguments{
\item{mat.z}{A matrix object of time series, regressor matrix}

\item{mat.y.ex}{A matrix object of time series, regressor matrix}

\item{n.eq}{number of equations in the VAR}

\item{p}{number of observations}

\item{est.mode}{estimation mode: "OLS", "FGLS", or "IGLS"}
}
\value{
A list with the matrix of beta parameters as first element and the covariance matrix of error as second element.
}
\description{
#Compute the matrix of parameters and the covariance matrix of errors in OLS, FGLS, or IGLS mode.
}
