% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{ConfidenceInterval}
\alias{ConfidenceInterval}
\title{ConfidenceInterval}
\usage{
ConfidenceInterval(mat.g, mat.s, mat.sigma, mat.r, mat.beta, cv, p)
}
\arguments{
\item{mat.g}{A matrix object of time series}

\item{mat.s}{A selection matrix}

\item{mat.sigma}{The covariance matrix}

\item{mat.r}{The selection vector of parameters}

\item{mat.beta}{The matrix of parameters}

\item{cv}{A vector of critical values}

\item{p}{The length of the vector}
}
\value{
The difference in time unit around the break
}
\description{
Compute the confidence interval in time unit.
}
