% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformable_matrix.R
\name{ConformableMatrix}
\alias{ConformableMatrix}
\title{ConformableMatrix}
\usage{
ConformableMatrix(mat.y, q, mat.x = NULL, trend = FALSE, intercept = TRUE)
}
\arguments{
\item{mat.y}{The matrix object of time series}

\item{q}{The chosen lag}

\item{mat.x}{The matrix of optional covariates}

\item{trend}{Whether we add a trend. Default = FALSE}

\item{intercept}{Whether the break test is on the intercept only. Default = TRUE}
}
\value{
A list of conformed matrices
}
\description{
Compute the list of matrices with correct dimensions to pass later in the computation
}
\examples{
data(example_data)
conf.matrix <- ConformableMatrix(mat.y = example_data, q = 2)
}
