% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{Fstat}
\alias{Fstat}
\title{Fstat}
\usage{
Fstat(mat.r, mat.beta, mat.z, p, mat.sigma)
}
\arguments{
\item{mat.r}{The selection matrix for the parameters}

\item{mat.beta}{The matrix of parameters}

\item{mat.z}{The matrix of original and "breaking" time series}

\item{p}{The number of observations}

\item{mat.sigma}{The covariance matrix}
}
\value{
The f-statistic scalar
}
\description{
Compute the f-statistic for the break test
}
