% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lags.R
\name{Lags}
\alias{Lags}
\title{Lags}
\usage{
Lags(mat.y, q)
}
\arguments{
\item{mat.y}{The matrix of time series}

\item{q}{The lag chosen}
}
\value{
A list of original (dependent) and lagged (independent) time series matrix
}
\description{
Compute the lags for the mat.y time series matrix
}
\examples{
data(example_data)
list.lags <- Lags(mat.y = example_data, q = 2)
}
