% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stats.R
\name{PlotStats}
\alias{PlotStats}
\title{PlotStats}
\usage{
PlotStats(my.dates, my.vars, f.stat, mat.ci = mat.ci, mat.y)
}
\arguments{
\item{my.dates}{A vector of dates}

\item{my.vars}{The variables tested}

\item{f.stat}{The f-statistics}

\item{mat.ci}{The matrix of confidence intervals}

\item{mat.y}{The original time series}
}
\value{
A ggplot2 object
}
\description{
Generate a ggplot2 object to depict the break and the time series tested
}
