% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{Sigma}
\alias{Sigma}
\title{Sigma}
\usage{
Sigma(mat.z, mat.y.ex, mat.beta, n.eq)
}
\arguments{
\item{mat.z}{A matrix of breaking and non breaking time series}

\item{mat.y.ex}{A vectorized matrix of time series}

\item{mat.beta}{The matrix of parameters}

\item{n.eq}{The number of equations in the VAR system}
}
\value{
The covariance matrix of errors
}
\description{
#compute the covariance matrix of errors as in Bai, Lumsdaine, and Stock (1998)
}
