% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{Simul}
\alias{Simul}
\title{Simul}
\usage{
Simul(n = 100, p = 5, intensity = 1, when.break = 0.5)
}
\arguments{
\item{n}{The number of time series observations}

\item{p}{The number of time series}

\item{intensity}{The intensity of the break}

\item{when.break}{When should the break be simulated (as a percentage of the time series sample)}
}
\value{
A matrix of time series with a common break
}
\description{
#Simulate data to test the functions
}
\examples{
data(example_data)
simul.data <- Simul(n = 100, p = 5, intensity = 1, when.break = 0.5)
}
