% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_distrib.R
\name{Vdistr}
\alias{Vdistr}
\title{Vdistr}
\usage{
Vdistr(ci)
}
\arguments{
\item{ci}{A vector of confidence intervals}
}
\value{
A vector of critical values
}
\description{
Computes the critical values for a vector of confidence intervals proposed (ci)
}
